Move services/__init__.py to services/base.py
Separate the base class implementations (and their utility functions,
imports, etc) from the 'service' package's top-level namespace. This
should encourage developers to be more conscientious about changing the
"public" API and should make API users more conscientious about
importing "private" symbols.
